Refersh Token reissueCount 자료형 수정

+++ backend/orury/src/main/java/com/kernel360/orury/domain/user/db/RefreshTokenEntity.java
@@ -24,7 +24,7 @@ public class RefreshTokenEntity {
 	@JoinColumn(name = "user_id")
 	private UserEntity user;
 	private String refreshToken;
-	private int reissueCount = 0;
+	private long reissueCount = 0;
 
 	public RefreshTokenEntity(UserEntity user, String refreshToken) {
 		this.user = user;
